--- a/Makefile
+++ b/Makefile
@@ -1,0 +1,8 @@
+
+test:
+	@go test -cover .
+
+bench:
+	@go test -bench=. -benchmem
+
+.PHONY: bench test