before:

```
BenchmarkGauge-8    	  300000	      5244 ns/op	     240 B/op	      10 allocs/op
BenchmarkIncr-8     	  300000	      4886 ns/op	     208 B/op	      10 allocs/op
BenchmarkDecr-8     	  300000	      4957 ns/op	     208 B/op	      10 allocs/op
BenchmarkHistogram-8	  300000	      5241 ns/op	     304 B/op	      10 allocs/op
BenchmarkDuration-8 	  300000	      5250 ns/op	     240 B/op	      10 allocs/op
```

after:

```
BenchmarkGauge-8    	 2000000	       637 ns/op	     276 B/op	       3 allocs/op
BenchmarkIncr-8     	 2000000	       620 ns/op	     210 B/op	       3 allocs/op
BenchmarkDecr-8     	 2000000	       629 ns/op	     208 B/op	       3 allocs/op
BenchmarkHistogram-8	 2000000	       633 ns/op	     276 B/op	       3 allocs/op
BenchmarkDuration-8 	 2000000	       641 ns/op	     276 B/op	       3 allocs/op
BenchmarkUnique-8   	 3000000	       537 ns/op	     216 B/op	       2 allocs/op
```